UI가 화면 크기에 따라 가변되도록 설정 UI화면은 640x1136 기준 Hero FSM구조로 변경 Live와 Dead 두가지 상태로 나눔 데미지 변수를 uint 에서 long으로 변경 10경 이상 표시가능 HeroSkill 시스템 구축(공속증가, 데미지증가, 탭 데미지증가 3가지 구현 중)

+++ TapperHeroes/Assets/Scripts/DamageText.cs
@@ -10,7 +10,7 @@ public class DamageText : MonoBehaviour
     Camera UICam = null;
     Vector3 Target = Vector3.zero;
     
-    public void Init(uint _Damage, Vector3 _Target, Color _Color, int _FontSize = 40, float _UpSpeed = 0.5f)
+    public void Init(long _Damage, Vector3 _Target, Color _Color, int _FontSize = 40, float _UpSpeed = 0.5f)
     {
         Target = _Target;
         Label = GetComponent<UILabel>();
